split up all tests they all run and pass now
